# openapi.yaml
openapi: 3.0.0
info:
  title: User Authentication and Account API
  version: 1.0.0
  description: API for user sign-up, sign-in, email verification, password management, and session handling.
servers:
  - url: /api # Assuming your API routes are under /api
components:
  schemas:
...
    Error:
      type: object
      properties:
        message:
          type: string
          description: A human-readable error message.
        details: # Optional, for more specific error info
          type: string
          nullable: true
...
    FeatureResponse: # Renamed from Feature and added id
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the feature.
        name:
          type: string
          description: Name of the feature.
        included:
          type: boolean
          description: Whether the feature is included.
      required:
        - id
        - name
        - included
    FeatureInput:
      type: object
      properties:
        id:
          type: string
          nullable: true
          description: Unique identifier for the feature (used for updates, optional for new features within a package update).
        name:
          type: string
          description: Name of the feature.
        included:
          type: boolean
          description: Whether the feature is included.
      required:
        - name
        - included
    CategoryBrief:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the category.
        name:
          type: string
          description: Name of the category.
        icon:
          type: string
          nullable: true
          description: Icon for the category (e.g., SVG content or a class name).
      required:
        - id
        - name
    SubcategoryBrief:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the subcategory.
        name:
          type: string
          description: Name of the subcategory.
        categoryId:
          type: string
          description: ID of the parent category.
      required:
        - id
        - name
        - categoryId
    AddonBrief:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
          format: float
        image:
          type: string
          format: url
          nullable: true
        categoryId:
          type: string
      required:
        - id
        - name
        - price
        - categoryId
    PackageBrief:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
          format: float
        image:
          type: string
          format: url
          nullable: true
        categoryId:
          type: string
        subcategoryId:
          type: string
        popular:
          type: boolean
          nullable: true
        bgColor:
          type: string
          nullable: true
      required:
        - id
        - name
        - price
        - categoryId
        - subcategoryId
    Category: # Updated to be CategoryFullResponse
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        icon:
          type: string
          nullable: true
        subcategories:
          type: array
          items:
            $ref: '#/components/schemas/SubcategoryBrief'
        addons:
          type: array
          items:
            $ref: '#/components/schemas/AddonBrief'
        packages:
          type: array
          items:
            $ref: '#/components/schemas/PackageBrief'
      required:
        - id
        - name
    CreateCategoryRequest:
      type: object
      properties:
        name:
          type: string
          description: Name of the category.
        icon:
          type: string
          nullable: true
          description: Icon for the category.
      required:
        - name
...
    Package: # Updated to be PackageFullResponse
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
          format: float
        image:
          type: string
          format: url
          nullable: true
        categoryId:
          type: string
        subcategoryId:
          type: string
        popular:
          type: boolean
          nullable: true
        bgColor:
          type: string
          nullable: true
        features:
          type: array
          items:
            $ref: '#/components/schemas/FeatureResponse'
        category:
          $ref: '#/components/schemas/CategoryBrief'
        subcategory:
          $ref: '#/components/schemas/SubcategoryBrief'
      required:
        - id
        - name
        - price
        - categoryId
        - subcategoryId
...
    UpdatePackageRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
          format: float
        image:
          type: string
          format: url
          nullable: true
        categoryId:
          type: string
        subcategoryId:
          type: string
        popular:
          type: boolean
        bgColor:
          type: string
          nullable: true # Allow explicitly setting to null
        features:
          type: array
          items:
            $ref: '#/components/schemas/FeatureInput' # id can be used to update existing, or omitted for new
          description: List of features to update/replace for the package.
  securitySchemes:
    cookieAuth: # For NextAuth.js session cookies
      type: apiKey
      in: cookie
      name: next-auth.session-token # Or __Secure-next-auth.session-token for HTTPS
security:
  - cookieAuth: [] # Indicates that most authenticated routes use cookieAuth
paths:
...
  /product/categories:
    post:
      summary: Create a new category
      tags: [Products]
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCategoryRequest'
      responses:
        '201':
          description: Category created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category' # Returns the full category object
        '400':
          description: Invalid input (e.g., name missing, validation error).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Conflict (e.g., category with this name already exists).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      summary: Get all categories
      tags: [Products]
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Successfully retrieved all categories.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category' # Returns full category objects with relations
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /product/packages/{packageId}:
    parameters:
      - name: packageId
        in: path
        required: true
        description: ID of the package.
        schema:
          type: string
    get:
      summary: Get a specific package by ID
      tags: [Products]
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Successfully retrieved the package.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Package' # Returns full package with features
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Package not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Update a package
      tags: [Products]
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePackageRequest'
      responses:
        '200':
          description: Package updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Package' # Returns updated full package with features
        '400':
          description: Invalid input (e.g., category/subcategory mismatch, name conflict).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Package, Category, or Subcategory not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Conflict (e.g., package with this name already exists).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete a package
      tags: [Products]
      security:
        - cookieAuth: []
      responses:
        '204':
          description: Package deleted successfully.
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Package not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /product/images/upload:
    post:
      summary: Upload a product image
      tags: [Products]
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: The image file to upload (jpeg, png, gif, webp). Max 5MB.
      responses:
        '200': # Changed from 201 to 200 as it's returning data about an existing resource (the file path)
          description: File uploaded successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: File uploaded successfully
                  filePath:
                    type: string
                    format: uri-relative
                    example: /product-images/xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx.png
                    description: The public relative path to the uploaded image.
        '400':
          description: Bad request (e.g., no file, invalid file type, file too large).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error during file upload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...